Keep Ternary Operators on Single Line
This might be dumb edit, but I think keeping the ternary operators on a single line makes it a lot more readable. I thought about throwing in some extra parens as well to make it more obvious that you're assigning the properties of each resulting object into the return object and making the `...!!` a little more distinct as _two_ operations, because as an intermediate dev this threw me for a minutes. 

`...(!!meta ? { meta } : {}),`

vs. 

`...!!meta ? { meta } : {},`

Not sure if 'readability for average joes' is something you'd care about spending a few extra characters on though.